pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["lingxt <lxt9601@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.50.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 99
